Fix for hanging splash screen
The splash screen hangs and then closes after a wait.

Observations

- The powershell command can fail but not terminate
- I the process hasn't terminated, the timeout kicks in but the streams
can't be closed. This causes the application to hang when trying to
close either the output or the error stream in the try-with-recourses

Two changes have been made:

* Added  -ErrorAction Stop to the powershell command which causes the
process to stop on the first error. This lets the process return without
needing to wait for the timeout
* Added a check to see if the process is still active and then destroy
if it is

Note: Should this check only be run in the the timout block?